🔨 Refactor(#79): sectionQueries 지역성 원칙 준수
