+++ tsconfig.app.tsbuildinfo
@@ -1,0 +1,1 @@
+{"root":["./src/app.tsx","./src/modalportal.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/axios/instance.ts","./src/axios/interceptor.ts","./src/axios/statuscode.ts","./src/common/layout/cokologo.tsx","./src/common/layout/header.tsx","./src/common/layout/menubar.tsx","./src/common/layout/modal.tsx","./src/common/layout/style.ts","./src/common/ui/headeritem.tsx","./src/common/ui/menuitem.tsx","./src/common/ui/style.ts","./src/features/learn/apis.ts","./src/features/learn/types.ts","./src/features/learn/service/getpartgridposition.ts","./src/features/learn/ui/keycapadventureintro.tsx","./src/features/learn/ui/sectionarrowbutton.tsx","./src/features/learn/ui/sectionnavigatecontainer.tsx","./src/features/learn/ui/selectsection.tsx","./src/features/learn/ui/style.ts","./src/features/login/styles.ts","./src/features/login/service/handlelogin.ts","./src/features/login/service/handlelogout.ts","./src/features/login/ui/login.tsx","./src/features/login/ui/loginprompt.tsx","./src/features/progress/style.ts","./src/features/progress/ui/progressbar.tsx","./src/features/quest/style.ts","./src/features/quest/ui/dailyquest.tsx","./src/features/quest/ui/mainquset.tsx","./src/features/quiz/apis.ts","./src/features/quiz/queries.ts","./src/features/quiz/types.ts","./src/features/quiz/service/addlinenumberstocode.ts","./src/features/quiz/service/replaceemptywithhtmlelement.ts","./src/features/quiz/service/usecodehighlight.ts","./src/features/quiz/ui/combination.tsx","./src/features/quiz/ui/multiplechoice.tsx","./src/features/quiz/ui/oxselector.tsx","./src/features/quiz/ui/partclear.tsx","./src/features/quiz/ui/partnavcontainer.tsx","./src/features/quiz/ui/question.tsx","./src/features/quiz/ui/quizsection.tsx","./src/features/quiz/ui/result.tsx","./src/features/quiz/ui/shortanswer.tsx","./src/features/quiz/ui/textblock.tsx","./src/features/quiz/ui/totalresults.tsx","./src/features/quiz/ui/styles.ts","./src/features/ranking/constant.ts","./src/features/ranking/ui/myrank.tsx","./src/features/ranking/ui/rankingcontainer.tsx","./src/features/ranking/ui/sortdropdown.tsx","./src/features/ranking/ui/styles.ts","./src/features/store/types.ts","./src/features/store/ui/cartlist.tsx","./src/features/store/ui/itemcontainer.tsx","./src/features/store/ui/storeitem.tsx","./src/features/store/ui/styles.ts","./src/features/user/apis.ts","./src/features/user/queries.ts","./src/features/user/types.ts","./src/features/user/ui/badgecontainer.tsx","./src/features/user/ui/mycharacter.tsx","./src/features/user/ui/profileimage.tsx","./src/features/user/ui/styles.ts","./src/hooks/usebeforeunload.ts","./src/hooks/usefunnel.tsx","./src/hooks/usemodal.ts","./src/hooks/useoutsideclick.ts","./src/hooks/usepopover.ts","./src/hooks/usepreloadimages.ts","./src/hooks/usescrollvisibility.ts","./src/pages/learn/learn.tsx","./src/pages/learn/style.ts","./src/pages/profile/profile.tsx","./src/pages/profile/styles.ts","./src/pages/quest/quest.tsx","./src/pages/quiz/quiz.tsx","./src/pages/quiz/styles.ts","./src/pages/ranking/ranking.tsx","./src/pages/ranking/styles.ts","./src/pages/store/store.tsx","./src/pages/store/styles.ts","./src/queries/sectionsquery.ts","./src/route/router.tsx","./src/store/useclientquizstore.ts","./src/store/usedndstore.ts","./src/store/useuserstore.ts","./src/style/globalstyle.ts","./src/style/layout.ts","./src/style/media.ts","./src/style/style.ts","./src/types/layout.ts","./src/utils/compact.ts","./src/utils/componentmap.tsx","./src/utils/cookies.ts","./src/utils/getimageurl.ts","./src/utils/handlepage.ts","./src/utils/isequalarray.ts","./src/utils/isloggedin.ts"],"errors":true,"version":"5.6.2"}
